fix!: fixed order of operations to match rust's
BREAKING CHANGE:
The order of operations has changed
to match what Rust does.
Previously, statements were evaluated strictly left to right.
While this new behavior is likely what was already expected,
it is _not_ what was previously done,
so I'm marking it as a breaking change
to be clear about that change.